+++ Pandora/Core.hs
@@ -1,4 +1,3 @@
 module Pandora.Core (module Exports) where
 
-import Pandora.Core.Morphism as Exports
 import Pandora.Core.Functor as Exports